added claim approval

+++ insurance-service/src/main/java/com/mtit/microservice/documentservice/documentservice/repository/ClaimRepositroy.java
@@ -3,6 +3,6 @@ package com.mtit.microservice.documentservice.documentservice.repository;
 import com.mtit.microservice.documentservice.documentservice.util.Claim;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ClaimRepositroy extends JpaRepository<Claim, Long> {
+public interface ClaimRepositroy extends JpaRepository<Claim, Integer> {
 
 }
